pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "leonardo-lima-98",email = "leonardodelima98@outlook.com"}
]
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "msal (>=1.33.0,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
lint = 'ruff check .'
post_lint = 'ruff check . --diff'
pre_format = 'ruff check . --fix'
format = ' ruff format .'
run = 'fastapi dev src/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html ; python -m http.server 8090 --directory htmlcov/'
autocommit = '. ./scripts/autocommit.sh'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.ruff]
line-length = 92
extend-exclude = ['migrations', 'pyproject.toml']

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
